% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dicom_to_nifti.R
\name{convert_dicom_to_nifti_R}
\alias{convert_dicom_to_nifti_R}
\title{Convert DICOM to NIFTI}
\usage{
convert_dicom_to_nifti_R(path)
}
\arguments{
\item{paths}{Output of \code{process_single_scan_R}.}
}
\description{
This function converts a DICOM image and its mask to the NIFTI format.
}
