% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{calculate_radiomic_features_R}
\alias{calculate_radiomic_features_R}
\title{Calculate Radiomic Features}
\usage{
calculate_radiomic_features_R(
  path_to_original_image,
  path_to_mask,
  path_to_parameters
)
}
\arguments{
\item{path_to_original_image}{A character string specifying the path to the original image.}

\item{path_to_mask}{A character string specifying the path to the image mask.}

\item{path_to_parameters}{A character string specifying the path to the parameter file for feature extraction.}
}
\value{
A list of computed radiomic features.
}
\description{
This function calculates radiomic features based on a given original image,
its mask, and a set of parameters.
}
\examples{
\dontrun{
features <- calculate_radiomic_features_R("/path/to/original/image", "/path/to/mask", "/path/to/parameters")
print(features)
}
}
